Bruker hook useHentOppgaver i OppgavebenkSide. Lagt til ett nytt valg 'Ingen' i radio group. Validering på at noe er valgt og at år er satt hvis kontroll av selvstendig er valgt. Flyttet kode for ferdigstilling av fremleggsoppgaver over i ny komponent.
